fix(v3.0.0): 修复 TwoPi 常量 double→int 隐式转换导致 CI 编译失败
MainWindow.xaml.cs 453 行 const int TwoPi = Math.PI*2.0 改为 const double,
消除 CS0266 隐式转换错误。

+++ src/MemoryBlackHole/Views/MainWindow.xaml.cs
@@ -450,7 +450,7 @@ namespace MemoryBlackHole.Views
         private sealed class SpaceCore
         {
             private const int OrbitPoolSize = 60;
-            private const int TwoPi = Math.PI * 2.0;
+            private const double TwoPi = Math.PI * 2.0;
 
             // 整体尺寸缩放系数(v3.0.1):1.0 是基线,放大后 BlurEffect 仍会糊。
             // 用户在第 4 轮反馈"放大 1.5x 后还是很糊",这里固定 1.0,
